requirements.txt
pdfplumber>=0.7.6
python-docx>=0.8.11
Flask>=2.0
pytest>=7.0